requirements.txt
dash
dash-bootstrap-components
pandas
pandas-stubs
plotly
pyvo
scikit-image
astroquery
astropy
aplpy
numexpr
matplotlib_venn
dash_aladin-lite
seaborn
coverage
selenium
dash[diskcache]
flask_login